refactor: extract Clock singleton
