+++ src/main/java/com/holub/model/Universe.java
@@ -1,0 +1,5 @@
+package com.holub.model;
+
+public class Universe {
+
+}
